+++ dotenv/load.ts
@@ -1,5 +1,5 @@
 // Copyright 2018-2022 the Deno authors. All rights reserved. MIT license.
 
-import { config } from "./mod.ts";
+import { configSync } from "./mod.ts";
 
-await config({ export: true });
+configSync({ export: true });