+++ node/internal/readline/interface.mjs
@@ -43,6 +43,7 @@ import {
   kSubstringSearch,
 } from "./utils.mjs";
 import { clearScreenDown, cursorTo, moveCursor } from "./callbacks.mjs";
+import { Readable } from "../../_stream.mjs";
 
 import { StringDecoder } from "../../string_decoder.ts";
 import {
@@ -78,9 +79,6 @@ import {
   kWordRight,
   kWriteToOutput,
 } from "./symbols.mjs";
-
-// Lazy load Readable for startup performance.
-let Readable;
 
 const kHistorySize = 30;
 const kMincrlfDelay = 100;
@@ -1186,9 +1184,6 @@ export class Interface extends InterfaceConstructor {
    */
   [Symbol.asyncIterator]() {
     if (this[kLineObjectStream] === undefined) {
-      if (Readable === undefined) {
-        Readable = require("stream").Readable;
-      }
       const readable = new Readable({
         objectMode: true,
         read: () => {